@RazrFalcon 2020-04-01
I guess there is nothing that stops us, but could you show an error with `Rc`? I'm interested in what exactly it doesn't like about `Rc`.