@tomjw64 2020-04-01
Here is a snippet of the error:
```
error[E0277]: `std::rc::Rc<std::string::String>` cannot be shared between threads safely
  --> src/main.rs:10:29
   |
10 |     (0..1000).into_par_iter().for_each(|_| {
   |  _____________________________^^^^^^^^_-
   | |                             |
   | |                             `std::rc::Rc<std::string::String>` cannot be shared between threads safely
11 | |     root.descendants().filter(|node| {
12 | |       node.tag_name().name() == "p"
13 | |     }).count();
14 | |   });
   | |___- within this `[closure@src/main.rs:10:38: 14:4 root:&roxmltree::Node<'_, '_>]`
   |
   = help: within `[closure@src/main.rs:10:38: 14:4 root:&roxmltree::Node<'_, '_>]`, the trait `std::marker::Sync` is not implemented for `std::rc::Rc<std::string::String>`
   = note: required because it appears within the type `roxmltree::Uri`
```