@RazrFalcon 2020-04-01
I've ran the bechmarks and it's 5-10% slower, which is pretty serious. I will think if there is another way, since we don't have a mutable access there is no need in `Arc`.