@tomjw64 2020-04-01
Curious. I must be doing something wrong while I'm benchmarking. Here were my results for completeness sake.
Before (w/Rc):
```
test large_roxmltree     ... bench:   4,990,875 ns/iter (+/- 372,508)
test medium_roxmltree    ... bench:     980,338 ns/iter (+/- 25,299)
test tiny_roxmltree      ... bench:       4,890 ns/iter (+/- 237)
```
After (w/Arc):
```
test large_roxmltree     ... bench:   4,753,714 ns/iter (+/- 149,681)
test medium_roxmltree    ... bench:     922,787 ns/iter (+/- 216,640)
test tiny_roxmltree      ... bench:       4,944 ns/iter (+/- 305)
```
I took it to mean that there was no difference, since everything was within the deviation. Any tips for making my benchmarking more reliable?

Good point that there should be no use for `Arc`. Let me see if I can whip anything up as well.