@RazrFalcon 2020-04-01
Yes, the measurements are a bit unstable. After multiple runs it looks like it can even be faster in some cases. 

The problem is that benches are testing only the parsing, but `Arc` will affect the public API itself, since every access should be atomic now.

I guess I will simply replace `Uri` with `Cow`. It makes `large_roxmltree` a bit slower, for whatever reason, but `medium_roxmltree` and `tiny_roxmltree` are faster than Rc/Arc now. So it's a fair tradeoff.

The only downside is that when an XML have a lot of not well-formed namespaces (e.g. it has `&\t\n\r`) than it will use way more memory. The idea behind `Uri` is to reduce the memory overhead of duplicated namespaces. But as long as we have a well-formed ns, we simply storing a `&str`.