@RazrFalcon 2020-04-01
I've pushed a new commit. Please try it out.