@tomjw64 2020-04-01
It works great! Hopefully you don't mind if I send in another PR here and there, even though I am probably less experienced. I have really enjoyed using this library and hope to give back! Thank you!