@tomjw64 2020-04-01
Closing in favor of https://github.com/RazrFalcon/roxmltree/commit/13fa4789584ada2f26cf37837a5b26e37f63cec1