@tomjw64 2020-04-01
@RazrFalcon I thought you might want to know, that with the change you pushed, I tried `cargo clippy` and it suggested the following:
```
warning: calling `to_string` on `&&mut std::borrow::Cow<'input, str>`
   --> src/parse.rs:798:49
    |
798 |                         *prev_text = Cow::Owned(prev_text.to_string() + text.borrow());
    |                                                 ^^^^^^^^^^^^^^^^^^^^^ help: try dereferencing the receiver: `(*prev_text).to_string()`
    |
    = note: `#[warn(clippy::inefficient_to_string)]` on by default
    = help: `&mut std::borrow::Cow<'input, str>` implements `ToString` through a slower blanket impl, but `std::borrow::Cow<'input, str>` has a fast specialization of `ToString`
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#inefficient_to_string
```
I'm sure it's not a big issue, but I thought I'd mention, just in case.