@RazrFalcon 2020-04-02
>Hopefully you don't mind if I send in another PR here and there

Sure. Any suggestions are always welcome. Just in this case, the simplest solution wasn't the right one.