@RazrFalcon 2020-04-02
I've fixed the clippy warnings. Thanks.